Moved parallel files into imageProcessing/parallel. Have yet to commit any .cu files.
